Move Aligner into a separate file
We need the aln_info available in a separate header file. It makes most
sense to arrange things so that everything that produces aln_info objects is
in aligner.cpp/.hpp
