+++ src/aln.hpp
@@ -1,5 +1,5 @@
-#ifndef aln_hpp
-#define aln_hpp
+#ifndef STROBEALIGN_ALN_HPP
+#define STROBEALIGN_ALN_HPP
 
 #include <string>
 #include <vector>
@@ -8,23 +8,7 @@
 #include "refs.hpp"
 #include "sam.hpp"
 #include "ssw_cpp.h"
-
-struct alignment_params {
-    // match is a score, the others are penalties (all are nonnegative)
-    int match;
-    int mismatch;
-    int gap_open;
-    int gap_extend;
-};
-
-struct aln_info {
-    std::string cigar;
-    unsigned int ed{0};  // edit distance
-    unsigned int ref_offset{0};
-    int sw_score{0};
-    int global_ed{0};  // edit distance plus total number of soft-clipped bases
-    int length{0};  // length of aligned reference segment
-};
+#include "aligner.hpp"
 
 struct AlignmentStatistics {
     std::chrono::duration<double> tot_read_file{0};
@@ -85,27 +69,6 @@ public:
     void update(int dist);
 };
 
-struct Aligner {
-public:
-    Aligner(alignment_params parameters)
-        : parameters(parameters)
-        , ssw_aligner(StripedSmithWaterman::Aligner(parameters.match, parameters.mismatch, parameters.gap_open, parameters.gap_extend))
-    { }
-
-    aln_info align(const std::string &ref, const std::string &query) const;
-
-    alignment_params parameters;
-
-    unsigned calls_count() {
-        return m_align_calls;
-    }
-
-private:
-    const StripedSmithWaterman::Aligner ssw_aligner;
-    const StripedSmithWaterman::Filter filter;
-    mutable unsigned m_align_calls{0};  // no. of calls to the align() method
-};
-
 void align_PE_read(
     const klibpp::KSeq& record1,
     const klibpp::KSeq& record2,
@@ -136,7 +99,6 @@ void align_SE_read(
 
 bool has_shared_substring(const std::string& read_seq, const std::string& ref_seq, int k);
 
-aln_info hamming_align(const std::string &query, const std::string &ref, int match, int mismatch, int &soft_left, int &soft_right);
 std::pair<size_t, size_t> highest_scoring_segment(const std::string& query, const std::string& ref, int match, int mismatch);
 
 #endif
